Switch to modular contact sensor framework
Each contact sensor manages its own contact state and dx during contact.
The contact detector may have an arbitrary number of contact sensors, by default
treated independently. For the bubble grippers, the bubbles are independent to
each other, but the residual contact sensor is not since it's likely that
if one of the bubble sensors detect a contact, the contact sensed by the
residual sensor refers to the same contact. Thus, the bubble sensors have
priority over the residual sensor, with the residual sensor reporting a contact
only if the bubble sensors do not detect a contact.